Moving url params logic to utils, keeping search drawer fixed on mobile
